chore(gatsby): migrate schema-composer and types/type-defs to typescript
